+++ src/TraderConfig.php
@@ -44,15 +44,6 @@ interface TraderConfig
     public function includeVatInPrices(): bool;
 
     /**
-     * Prices will be calculated including or excluding vat. This makes sure that calculations are correct
-     * and don't cause any rounding errors - which could occur when calculating excluding vat and
-     * including the vat afterwards. this can be set according to the visitor demands (b2b or b2c).
-     *
-     * @return bool
-     */
-    public function areItemPricesCalculatedExcludingVat(): bool;
-
-    /**
      * Calculate item discounts on the prices including or excluding vat.
      * This setting is mostly relevant for b2b shops where discounts are
      * often calculated on the prices excluding vat. Order discounts